add:FavoriteAndsearch_function

+++ resources/views/favorites/favorite_button.blade.php
@@ -1,17 +1,21 @@
-@if (Auth::id() != $user->id)
+@if(Auth::check())
 
-    @if (Auth::user()->is_favorite($song->id))
+    @if (Auth::id() != $user->id)
 
-        {!! Form::open(['route' => ['favorites.unfavorite', $song->id], 'method' => 'delete']) !!}
-            {!! Form::submit('いいね！を外す', ['class' => "button btn btn-warning"]) !!}
-        {!! Form::close() !!}
-
-    @else
-
-        {!! Form::open(['route' => ['favorites.favorite', $song->id]]) !!}
-            {!! Form::submit('いいね！を付ける', ['class' => "button btn btn-success"]) !!}
-        {!! Form::close() !!}
-
+        @if ($song)
+        
+            @if (Auth::user()->is_favorite($song->id))
+                {!! Form::open(['route' => ['favorites.unfavorite', $song->id], 'method' => 'delete']) !!}
+                    {!! Form::submit('いいね！外す', ['class' => "button btn btn-warning col-sm-6"]) !!}
+                {!! Form::close() !!}
+            @else
+                {!! Form::open(['route' => ['favorites.favorite', $song->id], 'method' => 'post']) !!}
+                    {!! Form::submit('いいね！', ['class' => "button btn btn-success col-sm-6"]) !!}
+                {!! Form::close() !!}
+            @endif
+            
+        @endif
+            
     @endif
-
+    
 @endif
